pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.1",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.123.8",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "oauthlib>=3.3.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pulp>=3.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
